ui improvements

+++ convex/validation.ts
@@ -11,7 +11,11 @@ export const gameQuickIdSchema = z
 
 export type GameQuickId = z.infer<typeof gameQuickIdSchema>;
 export const gameSecondsPerQuestionSchema = z.number().min(1).max(60);
-export const gameNumRoundsSchema = z.number().min(1).max(allQuestions.length);
+export const gameNumRoundsSchema = z
+  .number()
+  .int()
+  .min(1)
+  .max(allQuestions.length);
 export const gamePlayerGuessSchema = z.number().min(0).max(1);
 
 export const playerIdSchema = z.string().trim().length(10).brand("playerId");
